+++ lib/kakine.rb
@@ -11,7 +11,6 @@ require 'kakine/resource/openstack'
 require 'kakine/resource/yaml'
 require 'kakine/security_group'
 require 'kakine/security_rule'
-require 'kakine/validate'
 require 'kakine/errors'
 
 module Kakine
